requirements.txt
streamlit 
langchain 
langchain_ollama
langchain-openai
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv
playwright
cloudscraper
loguru
fastapi
uvicorn
pandas
openpyxl
apscheduler
pymongo
psycopg2-binary
redis
requests